Add UserLotto class to extends Lotto class
