+++ src/main/java/Lottos.java
@@ -3,15 +3,15 @@ import java.util.List;
 
 public class Lottos {
 	
-	List<Lotto> lottos;
+	List<UserLotto> lottos;
 	
-	Lottos(List<Lotto> lottos) {
+	Lottos(List<UserLotto> lottos) {
 		this.lottos = lottos;
 	}
 
 	List<Match> createMatches(List<Integer> winningLotto, int bonusNumber) {
 		List<Match> matches = new ArrayList<>();
-		for (Lotto eachLotto : this.lottos) {
+		for (UserLotto eachLotto : this.lottos) {
 			int matchingCount = eachLotto.makeMatchingCount(winningLotto);
 			boolean bonusTag = eachLotto.makeMatchingBonusTag(bonusNumber);
 			Match match = Match.findMatch(matchingCount, bonusTag);
