+++ src/main/java/Main.java
@@ -6,7 +6,7 @@ public class Main {
 		int howMany = Input.getHowMany(money);
 		int howManyManual = Input.getHowManyManual();
 		int howManyAuto = howMany - howManyManual;
-		Lottos lottos = new Lottos(Lotto.createLottos(howManyManual, howManyAuto));
+		Lottos lottos = new Lottos(UserLotto.createLottos(howManyManual, howManyAuto));
 		Result.showMyLottos(lottos);
 		WinningLotto winningLotto = new WinningLotto(WinningLotto.getWinningLotto());
 		int bonusNum = Input.getBonusNumber();
